backend/pyproject.toml
# backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Daniil <petrovdanil230@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.127.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.45"}
asyncpg = "^0.31.0"
bcrypt = "4.1.3"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
pydantic-settings = "^2.12.0"
alembic = "^1.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"